+++ src/app.rs
@@ -2,7 +2,6 @@ use crate::app::reputation_message::ReputationMessage;
 use crate::objects::messages::{MessageSender, ParseMessage};
 use crate::services::bootstrap;
 use handle_rep::HandledReputation;
-use message_data::MessageData;
 use std::future::Future;
 use teloxide::prelude::*;
 
@@ -10,7 +9,6 @@ pub mod handle_chat;
 pub mod handle_cmd;
 pub mod handle_rep;
 pub mod handle_user;
-pub mod message_data;
 pub mod reputation_message;
 
 pub fn init() -> impl Future {
@@ -20,20 +18,22 @@ pub fn init() -> impl Future {
     teloxide::repl(bot, |message: Message, bot: AutoSend<Bot>| async move {
         handle_user::save_user(&message);
         handle_chat::save_chat(&bot, &message).await;
-        let data = MessageData::get_data(&message);
-        let reputation = ReputationMessage::new(&message);
-        let (is_command, command_message) = handle_cmd::execute(&message);
+
+        let bot_username = handle_cmd::get_bot_username(&bot).await;
+
+        let (is_command, command_message) = handle_cmd::execute(&message, bot_username);
         if is_command {
             let command_message = command_message.unwrap_or("Unknown command error".to_string());
 
-            MessageSender::new(data.get_chat_id(), command_message).send(bot).await;
+            MessageSender::new(message.chat.id, command_message).send(bot).await;
             return respond(());
         }
 
+        let reputation = ReputationMessage::new(&message);
         let result = HandledReputation::handle_rep(&reputation);
         match result {
             Some(_handled_reputation) => {
-                MessageSender::new(data.get_chat_id(), _handled_reputation.parse())
+                MessageSender::new(message.chat.id, _handled_reputation.parse())
                     .send(bot)
                     .await;
             }